Fold _console.ts into _global.ts.
